+++ application/common/components/passwordStore/Multiple.php
@@ -8,7 +8,7 @@ use common\components\passwordStore\PasswordStoreInterface;
 use common\components\passwordStore\PasswordReuseException;
 use common\components\passwordStore\UserNotFoundException;
 use common\components\passwordStore\UserPasswordMeta;
-use common\components\passwordStore\NotAttemptedException;
+use common\components\passwordStore\PasswordStoreException;
 use yii\base\Component;
 
 class Multiple extends Component implements PasswordStoreInterface
@@ -18,7 +18,10 @@ class Multiple extends Component implements PasswordStoreInterface
     
     /** @var PasswordStoreInterface[] */
     protected $passwordStores = [];
-    
+
+    public $displayName = 'Multiple';
+
+
     /**
      * See if all the password store backends are available.
      *
@@ -36,19 +39,19 @@ class Multiple extends Component implements PasswordStoreInterface
             try {
                 $passwordStore->getMeta($employeeId);
             } catch (Exception $e) {
-                throw new NotAttemptedException(sprintf(
+                throw new \Exception(sprintf(
                     'Did not attempt to %s because not all of the backends are '
                     . 'available. The %s password store gave this error when '
                     . 'asked for the specified user (%s): %s',
                     $taskDescription,
-                    \get_class($passwordStore),
+                    $passwordStore->getDisplayName(),
                     var_export($employeeId, true),
                     $e->getMessage()
                 ), 1498163919, $e);
             }
         }
     }
-    
+
     public function init()
     {
         parent::init();
@@ -91,8 +94,10 @@ class Multiple extends Component implements PasswordStoreInterface
      * See if all of the password stores seem to be available/responding, and if
      * so set the user's password in all of the defined password stores. If any
      * of the password stores fail the "pre-check", this will not attempt to set
-     * the user's password on any of them, instead throwing a
-     * NotAttemptedException.
+     * the user's password on any of them, instead throwing a PasswordStoreException
+     * Thereafter, if any of the password stores fail, a PasswordStoreException will
+     * be thrown with a message detailing which ones succeeded and which ones
+     * failed.
      *
      * NOTE: If successful, this will return the UserPasswordMeta returned by
      *       the first password store defined in its list.
@@ -100,7 +105,7 @@ class Multiple extends Component implements PasswordStoreInterface
      * @param string $employeeId The Employee ID of the user.
      * @param string $password The new password.
      * @return UserPasswordMeta
-     * @throws NotAttemptedException
+     * @throws PasswordStoreException
      * @throws Exception
      * @throws UserNotFoundException
      * @throws AccountLockedException
@@ -109,27 +114,46 @@ class Multiple extends Component implements PasswordStoreInterface
     public function set($employeeId, $password): UserPasswordMeta
     {
         $this->assertAllBackendsAreAvailable($employeeId, 'set the password');
-        $numSuccessfullySet = 0;
+
         $responses = [];
+        $successes = [];
+        $errors = [];
         foreach ($this->passwordStores as $passwordStore) {
             try {
                 $responses[] = $passwordStore->set($employeeId, $password);
-                $numSuccessfullySet++;
+                $successes[] = $passwordStore->getDisplayName();
             } catch (Exception $e) {
                 if ($e instanceof PasswordReuseException) {
                     // Be aware that this does not include information about how many backends the password
                     // was successfully changed in
                     throw $e;
                 }
-                throw new Exception(sprintf(
-                    'Failed to set the password using %s after successfully '
-                    . 'setting it in %s other password store(s). Error: %s',
-                    \get_class($passwordStore),
-                    $numSuccessfullySet,
-                    $e->getMessage()
-                ), 1498162884, $e);
+
+                \Yii::error([
+                    'action' => 'set password',
+                    'status' => 'error',
+                    'passwordStore' => $passwordStore->getDisplayName(),
+                    'message' => $e->getMessage(),
+                ]);
+                $errors[] = $passwordStore->getDisplayName();
             }
         }
+
+        if (count($errors) > 0) {
+            $conjunction = ' ' . \Yii::t('app', 'and') . ' ';
+            $errorMessage = \Yii::t(
+                'app',
+                count($successes) > 0 ? 'Multiple.SetPartialSuccess' : 'Multiple.SetFailed',
+                [
+                        'successes' => implode($conjunction, $successes),
+                        'errors' => implode($conjunction, $errors),
+                        'supportName' => \Yii::$app->params['support']['name'],
+                        'supportEmail' => \Yii::$app->params['support']['email'],
+                ]
+            );
+            throw new PasswordStoreException($errorMessage, 1498162884);
+        }
+
         return $responses[0];
     }
 
@@ -164,5 +188,13 @@ class Multiple extends Component implements PasswordStoreInterface
             }
         }
         return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDisplayName(): string
+    {
+        return $this->displayName;
     }
 }